LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-sub-frame-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Cscript%3Eonload%20=%20function()%20{%20parent.postMessage('LOADED',%20'*');%20}%20%3C/script%3E%3Cbody%3E%3Cp%20id='accessMe'%3E%3C/p%3E%3Cp%3EInner%20iframe.%3C/p%3E%3C/body%3E%3C/html%3E from frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-to-data-url-sub-frame.html. Domains, protocols and ports must match.

This tests that the main frame can't access the contents of an iframe that contains a data: URL loaded page


PASS: Cross frame access to a data: URL was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.
